medical_insurance_project_v2/requirements.txt
# Core
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.1
scipy==1.13.1
joblib==1.4.2

# Models
xgboost==2.1.1
lightgbm==4.5.0

# Tracking
mlflow==2.16.0

# UI
streamlit==1.38.0
plotly==5.24.1
matplotlib==3.9.2
seaborn==0.13.2

# Reports
fpdf2==2.7.9
jinja2==3.1.4

# Validation / utils
pydantic==2.9.2
python-dotenv==1.0.1
PyYAML==6.0.2
